Move config loading funcs to their own file
