+++ go.mod
@@ -9,11 +9,10 @@ require (
 	github.com/patrickmn/go-cache v2.1.0+incompatible
 	github.com/samber/mo v1.11.0
 	github.com/stretchr/testify v1.9.0
-	gopkg.in/yaml.v2 v2.4.0
+	gopkg.in/yaml.v3 v3.0.1
 )
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
-	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
